Use memcpy() rather than stpncpy().
I've also switched a strcpy() to a memcpy() to make it obviously
optimal, even though clang is clever enough to compile either a strcpy()
or a memcpy() of a string constant into just two instructions: a 16-byte
load and a 16-byte store.

Since this is a hotspot, I've removed the implicit memset() for the
buffers too, and I'm using __builtin_memcpy() because clang will
otherwise call __memcpy_chk(), apparently because it's afraid that the
addition to calculate the VLA size could overflow.

Change-Id: Iba8a93c1a23611446d9b3faf4630b15d7455a658