Reland "Remove some C cruft."
This reverts commit 1ce2ae5769da2d137fd4b2d0f6d23d44525de695, adding
tests for comma-separated strings which are used in several CTS tests,
and switching to libbase Split() which takes a set of delimiter chatacters
rather than a single [possible multi-character] delimiter.

I was also going to build the logcat tests for the host (because
treehugger can't be trusted to actually run the current version of the
tests at the moment), but it turns out that treehugger will automatically
run even _new_ host tests in presubmit, and since most of these tests are
"exec /system/bin/logcat", all the tests except for the ones I've just
added fail. This still seems worth coming back to, but "not today".

Change-Id: I91fbb1217e6539d4280206e25491d355be2b686d