logd.auditd: Improve debuggability of avc denials with hex string paths
When generating avc denials, if the kernel detects a space or control
character in a file path, it will log the file path using a hex string
representing each charcter in the path instead of the actual path

For example:

... type=1400 audit(0.0:4399): avc:  denied  { read write } for
path=2F6D656D66643A63725F72656C726F202864656C6574656429 ...

The kernel emits the hex string representation of the file name in
these scenarios to avoid polluting the logs as explained in [1].

This is rather cryptic and unclear as to what the avc denial is related
to.

To make these types of avc denials easier to debug, detect when such an
avc denial occurs, and emit the decoded version of the path after
emitting the denial.

This ends up logging the following messages:

... type=1400 audit(0.0:50): avc:  denied  { read write } for
path=2F6D656D66643A63725F72656C726F202864656C6574656429 ...

... Decoded path for audit(0.0:50): /memfd:MessageQueue (deleted)

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=83c7d09

Bug: 439085655
Change-Id: Idc511619fcffaa06f8642c7eee6bfa92d75cb350