logd.auditd: Fix logspam from HexToBytes()
The logic that was introduced to parse hex strings from avc denials
calls HexToBytes() on a path as long as it's not empty. This is not
appropriate, as the path may already be a proper path and not an
encoded path.

When an avc denial is emitted and the path does not need to be encoded,
the kernel does enclose the path in quotes, so only invoke HexToBytes()
on a path if it does not start with quotes.

Bug: 439085655
Bug: 445461069
Change-Id: I23639d37ba0e49039afc7f16868d238047673c4a