Enable AFDO optimization for logd
A lot of CPU cycles are spent in logd during system boot, making it a
good target to enable AutoFDO.

This change asks the compiler to emit additional debugging info for
producing AutoFDO profiles, which can be used to enable profile guided
optimization of logd. We will conduct experiments to confirm the
performance difference.

Test: m logd
Bug: 453894736
Change-Id: I61d3bdb9d4ff85c0a56f21848c370356cb0993aa